create & findAll working

+++ src/main/scala/models/package.scala
@@ -2,5 +2,5 @@ package object models {
   type UserId = Long
   type PostId = Long
   type CommentId = Long
-  type EntityId = Long
+  type EntityId = Int
 }
